[+] refactor : modal component
